# Fixed Series.apply performance regression