--- a/pandas/core/apply.py
+++ b/pandas/core/apply.py
@@ -295,11 +295,10 @@ class FrameApply(metaclass=abc.ABCMeta):
                 for i, v in series_gen_enumeration:
                     # ignore SettingWithCopy here in case the user mutates
                     results[i] = self.f(v)
-
-                if isinstance(results[i], ABCSeries):
-                    # If we have a view on v, we need to make a copy because
-                    #  series_generator will swap out the underlying data
-                    results[i] = results[i].copy(deep=False)
+                    if isinstance(results[i], ABCSeries):
+                        # If we have a view on v, we need to make a copy because
+                        #  series_generator will swap out the underlying data
+                        results[i] = results[i].copy(deep=False)
 
         return results, res_index