Set the option once, rather than in the loop.

Closes https://github.com/pandas-dev/pandas/issues/35047


```python
import pandas as pd
import numpy as np

df = pd.DataFrame(np.random.randn(1000, 3), columns=list("ABC"))
%timeit df.apply(lambda x: x["A"] + x["B"], axis=1)
```

```
# 1.0.4
22.5 ms ± 4.73 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

# PR
10.4 ms ± 714 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```